generator/pyproject.toml
[tool.poetry]
name = "feed"
version = "0.1.0"
description = ""
authors = ["Julien Lemaitre <jlnait3@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
mutagen = "^1.47.0"
pyyaml = "^6.0.2"
requests = "^2.32.3"
jinja2 = "^3.1.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
